First step to refactor engine
