+++ ignite/engine/engine.py
@@ -3,251 +3,15 @@ import logging
 import time
 from collections import defaultdict, OrderedDict
 from collections.abc import Mapping
-from enum import Enum
 import weakref
-import numbers
 import random
 import warnings
 
 import torch
 
+from ignite.engine.events import Events, State, EventWithFilter, RemovableEventHandle
+from ignite.engine.utils import ReproducibleBatchSampler, _update_dataloader
 from ignite._utils import _to_hours_mins_secs
-
-
-class EventWithFilter:
-
-    def __init__(self, event, filter):
-        if not callable(filter):
-            raise TypeError("Argument filter should be callable")
-        self.event = event
-        self.filter = filter
-
-    def __str__(self):
-        return "<%s event=%s, filter=%r>" % (self.__class__.__name__, self.event, self.filter)
-
-
-class CallableEvents:
-    """Base class for Events implementing call operator and storing event filter. This class should be inherited
-    for any custom events with event filtering feature:
-
-    .. code-block:: python
-
-        from ignite.engine.engine import CallableEvents
-
-        class CustomEvents(CallableEvents, Enum):
-            TEST_EVENT = "test_event"
-
-        engine = ...
-        engine.register_events(*CustomEvents, event_to_attr={CustomEvents.TEST_EVENT: "test_event"})
-
-        @engine.on(CustomEvents.TEST_EVENT(every=5))
-        def call_on_test_event_every(engine):
-            # do something
-
-    """
-    def __call__(self, event_filter=None, every=None, once=None):
-
-        if not((event_filter is not None) ^ (every is not None) ^ (once is not None)):
-            raise ValueError("Only one of the input arguments should be specified")
-
-        if (event_filter is not None) and not callable(event_filter):
-            raise TypeError("Argument event_filter should be a callable")
-
-        if (every is not None) and not (isinstance(every, numbers.Integral) and every > 0):
-            raise ValueError("Argument every should be integer and greater than zero")
-
-        if (once is not None) and not (isinstance(once, numbers.Integral) and once > 0):
-            raise ValueError("Argument every should be integer and positive")
-
-        if every is not None:
-            if every == 1:
-                # Just return the event itself
-                return self
-            event_filter = CallableEvents.every_event_filter(every)
-
-        if once is not None:
-            event_filter = CallableEvents.once_event_filter(once)
-
-        # check signature:
-        Engine._check_signature("engine", event_filter, "event_filter", "event")
-
-        return EventWithFilter(self, event_filter)
-
-    @staticmethod
-    def every_event_filter(every):
-        def wrapper(engine, event):
-            if event % every == 0:
-                return True
-            return False
-        return wrapper
-
-    @staticmethod
-    def once_event_filter(once):
-        def wrapper(engine, event):
-            if event == once:
-                return True
-            return False
-        return wrapper
-
-
-class Events(CallableEvents, Enum):
-    """Events that are fired by the :class:`~ignite.engine.Engine` during execution.
-
-    Since v0.3.0, Events become more flexible and allow to pass an event filter to the Engine:
-
-    .. code-block:: python
-
-        engine = Engine()
-
-        # a) custom event filter
-        def custom_event_filter(engine, event):
-            if event in [1, 2, 5, 10, 50, 100]:
-                return True
-            return False
-
-        @engine.on(Events.ITERATION_STARTED(event_filter=custom_event_filter))
-        def call_on_special_event(engine):
-             # do something on 1, 2, 5, 10, 50, 100 iterations
-
-        # b) "every" event filter
-        @engine.on(Events.ITERATION_STARTED(every=10))
-        def call_every(engine):
-            # do something every 10th iteration
-
-        # c) "once" event filter
-        @engine.on(Events.ITERATION_STARTED(once=50))
-        def call_once(engine):
-            # do something on 50th iteration
-
-    Event filter function `event_filter` accepts as input `engine` and `event` and should return True/False.
-    Argument `event` is the value of iteration or epoch, depending on which type of Events the function is passed.
-
-    """
-    EPOCH_STARTED = "epoch_started"
-    EPOCH_COMPLETED = "epoch_completed"
-    STARTED = "started"
-    COMPLETED = "completed"
-    ITERATION_STARTED = "iteration_started"
-    ITERATION_COMPLETED = "iteration_completed"
-    EXCEPTION_RAISED = "exception_raised"
-
-    GET_BATCH_STARTED = "get_batch_started"
-    GET_BATCH_COMPLETED = "get_batch_completed"
-
-
-class State:
-    """An object that is used to pass internal and user-defined state between event handlers. By default, state
-    contains the following attributes:
-
-    .. code-block:: python
-
-        state.iteration         # 1-based, the first iteration is 1
-        state.epoch             # 1-based, the first epoch is 1
-        state.seed              # seed to set at each epoch
-        state.dataloader        # data passed to engine
-        state.epoch_length      # optional length of an epoch
-        state.max_epochs        # number of epochs to run
-        state.batch             # batch passed to `process_function`
-        state.output            # output of `process_function` after a single iteration
-        state.metrics           # dictionary with defined metrics if any
-
-    """
-
-    event_to_attr = {
-        Events.GET_BATCH_STARTED: "iteration",
-        Events.GET_BATCH_COMPLETED: "iteration",
-        Events.ITERATION_STARTED: "iteration",
-        Events.ITERATION_COMPLETED: "iteration",
-        Events.EPOCH_STARTED: "epoch",
-        Events.EPOCH_COMPLETED: "epoch",
-        Events.STARTED: "epoch",
-        Events.COMPLETED: "epoch",
-    }
-
-    def __init__(self, **kwargs):
-        self.iteration = 0
-        self.epoch = 0
-        self.epoch_length = None
-        self.max_epochs = None
-        self.output = None
-        self.batch = None
-        self.metrics = {}
-        self.dataloader = None
-        self.seed = None
-
-        for k, v in kwargs.items():
-            setattr(self, k, v)
-
-        for value in self.event_to_attr.values():
-            if not hasattr(self, value):
-                setattr(self, value, 0)
-
-    def get_event_attrib_value(self, event_name):
-        if isinstance(event_name, EventWithFilter):
-            event_name = event_name.event
-        if event_name not in State.event_to_attr:
-            raise RuntimeError("Unknown event name '{}'".format(event_name))
-        return getattr(self, State.event_to_attr[event_name])
-
-    def __repr__(self):
-        s = "State:\n"
-        for attr, value in self.__dict__.items():
-            if not isinstance(value, (numbers.Number, str)):
-                value = type(value)
-            s += "\t{}: {}\n".format(attr, value)
-        return s
-
-
-class RemovableEventHandle:
-    """A weakref handle to remove a registered event.
-
-    A handle that may be used to remove a registered event handler via the
-    remove method, with-statement, or context manager protocol. Returned from
-    :meth:`~ignite.engine.Engine.add_event_handler`.
-
-
-    Args:
-        event_name: Registered event name.
-        handler: Registered event handler, stored as weakref.
-        engine: Target engine, stored as weakref.
-
-    Example usage:
-
-    .. code-block:: python
-
-        engine = Engine()
-
-        def print_epoch(engine):
-            print("Epoch: {}".format(engine.state.epoch))
-
-        with engine.add_event_handler(Events.EPOCH_COMPLETED, print_epoch):
-            # print_epoch handler registered for a single run
-            engine.run(data)
-
-        # print_epoch handler is now unregistered
-    """
-
-    def __init__(self, event_name, handler, engine):
-        self.event_name = event_name
-        self.handler = weakref.ref(handler)
-        self.engine = weakref.ref(engine)
-
-    def remove(self):
-        """Remove handler from engine."""
-        handler = self.handler()
-        engine = self.engine()
-
-        if handler is None or engine is None:
-            return
-
-        if engine.has_event_handler(handler, self.event_name):
-            engine.remove_event_handler(handler, self.event_name)
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, type, value, tb):
-        self.remove()
 
 
 class Engine:
@@ -550,23 +314,6 @@ class Engine:
         if len(new_event_handlers) == len(self._event_handlers[event_name]):
             raise ValueError("Input handler '{}' is not found among registered event handlers".format(handler))
         self._event_handlers[event_name] = new_event_handlers
-
-    @staticmethod
-    def _check_signature(self, fn, fn_description, *args, **kwargs):
-        exception_msg = None
-
-        signature = inspect.signature(fn)
-        try:
-            signature.bind(self, *args, **kwargs)
-        except TypeError as exc:
-            fn_params = list(signature.parameters)
-            exception_msg = str(exc)
-
-        if exception_msg:
-            passed_params = [self] + list(args) + list(kwargs)
-            raise ValueError("Error adding {} '{}': "
-                             "takes parameters {} but will be called with {} "
-                             "({}).".format(fn, fn_description, fn_params, passed_params, exception_msg))
 
     def on(self, event_name, *args, **kwargs):
         """Decorator shortcut for add_event_handler.
@@ -959,51 +706,3 @@ class Engine:
 
         self._dataloader_iter = self._dataloader_len = None
         return self.state
-
-
-def _update_dataloader(dataloader, new_batch_sampler):
-    params_keys = [k for k in dataloader.__dict__.keys() if not k.startswith("_")]
-    for k in ['batch_size', 'sampler', 'drop_last', 'batch_sampler', 'dataset_kind']:
-        if k in params_keys:
-            params_keys.remove(k)
-    params = {k: getattr(dataloader, k) for k in params_keys}
-    params['batch_sampler'] = new_batch_sampler
-    return torch.utils.data.DataLoader(**params)
-
-
-class ReproducibleBatchSampler(torch.utils.data.sampler.BatchSampler):
-    """Reproducible batch sampler. Internally, this class iterates and stores indices of the input batch sampler.
-
-    Args:
-        batch_sampler (torch.utils.data.sampler.BatchSampler): batch sampler same as used with
-            `torch.utils.data.DataLoader`
-        start_iteration (int, optional): optional start iteration
-    """
-    def __init__(self, batch_sampler, start_iteration=None):
-        if not isinstance(batch_sampler, torch.utils.data.sampler.BatchSampler):
-            raise TypeError("Argument batch_sampler should be torch.utils.data.sampler.BatchSampler")
-
-        self.batch_indices = None
-        self.batch_sampler = batch_sampler
-        self.start_iteration = start_iteration
-        self.sampler = self.batch_sampler.sampler
-
-    def setup_batch_indices(self):
-        self.batch_indices = []
-        for batch in self.batch_sampler:
-            self.batch_indices.append(batch)
-
-        if self.start_iteration is not None:
-            self.batch_indices = self.batch_indices[self.start_iteration:]
-            self.start_iteration = None
-
-    def __iter__(self):
-        if self.batch_indices is None:
-            self.setup_batch_indices()
-        for batch in self.batch_indices:
-            yield batch
-
-        self.batch_indices = None
-
-    def __len__(self):
-        return len(self.batch_sampler)
